+++ src/store/dispatcher.mjs
@@ -8,11 +8,10 @@ const debug = createDebug( "tabulate:store:dispatcher" )
 export function createStore( reducer, initial_state ) {
   const listeners = []
   let current_state = initial_state
-  let is_dispatching = false
   let dispatch_id = 0
 
   const store = {
-    is_dispatching,
+    is_dispatching: false,
     dispatch,
     subscribe,
     getState,
@@ -26,13 +25,12 @@ export function createStore( reducer, initial_state ) {
     listeners.push( listener )
 
     return function unsubscribe() {
-      var index = listeners.indexOf( listener )
-      listeners.splice( index, 1 )
+      listeners.splice( listeners.indexOf( listener ), 1 )
     }
   }
 
   function dispatch( action ) {
-    let local_dispatch_id = ++dispatch_id
+    const local_dispatch_id = ++dispatch_id
     let is_unchanged = false
 
     try {