+++ test/store/reducers/group-add.mjs
@@ -7,14 +7,12 @@ import {
 import { addGroup } from "../../../src/store/reducers/group.mjs"
 import { validateState } from "../../../src/store/validators.mjs"
 
-function testNormalCreate( t ) {
+tap.test( function testNormalCreate( t ) {
   let state = getInitialState()
 
   let tab_id = state.windows[ 0 ].tab_groups[ 1 ].tabs[ 0 ].id
 
   state = addGroup( state, { tab_id, window_id: state.windows[ 0 ].id } )
-  t.ok( validateState( state ), "state validates", validateState.errors )
+  t.ok( validateState( state ), "should pass validation", validateState.errors )
   t.end()
-}
-
-tap.test( testNormalCreate )
+})