+++ test/store/reducers/group-move.mjs
@@ -15,22 +15,26 @@ import {
   createTestTab
 } from "../helpers.mjs"
 
-function toSameWindow( t ) {
+tap.test( function toSameWindow( t ) {
   const window_id = 1
   const tab_group_id = 3
   const tab_group_index = 1
   const state0 = {
     config: {},
     windows: [
-      createWindow( window_id, [
-        createPinnedTabGroup( [] ),
-        createTabGroup( 2, [
-          createTestTab({ id: 4 })
-        ]),
-        createTabGroup( tab_group_id, [
-          createTestTab({ id: 5 })
-        ])
-      ])
+      createWindow(
+        window_id,
+        [
+          createPinnedTabGroup( [] ),
+          createTabGroup( 2, [
+            createTestTab({ id: 4 })
+          ]),
+          createTabGroup( tab_group_id, [
+            createTestTab({ id: 5 })
+          ]),
+        ],
+        { active_tab_group_id: tab_group_id, active_tab_id: 5, highlighted_tab_ids: [ 5 ] },
+      )
     ]
   }
 
@@ -38,13 +42,15 @@ function toSameWindow( t ) {
   const target_data = { window_id, tab_group_index }
 
   const state1 = moveGroup( state0, { source_data, target_data } )
-  t.ok( validateState( state1 ), "state validates", validateState.errors )
+  t.ok( validateState( state1 ), "should pass validation", validateState.errors )
   t.equal( state1.windows[ 0 ].tab_groups[ tab_group_index ].id, tab_group_id )
+  t.equal( state1.windows[ 0 ].active_tab_group_id, tab_group_id, "should not update the active group during move" )
+  t.equal( state1.windows[ 0 ].active_tab_id, 5, "should not update the active tab during move" )
 
   t.end()
-}
+})
 
-function toSameWindowDropZone( t ) {
+tap.test( function toSameWindowDropZone( t ) {
   const window_id = 1
   const tab_group_id = 3
   const state0 = {
@@ -66,13 +72,13 @@ function toSameWindowDropZone( t ) {
   const target_data = { window_id, tab_group_index: null }
 
   const state1 = moveGroup( state0, { source_data, target_data } )
-  t.ok( validateState( state1 ), "state validates", validateState.errors )
+  t.ok( validateState( state1 ), "should pass validation", validateState.errors )
   t.equal( state1.windows[ 0 ].tab_groups[ 2 ].id, tab_group_id )
 
   t.end()
-}
+})
 
-function toDifferentWindow( t ) {
+tap.test( function toDifferentWindow( t ) {
   const tab_group_id = 3
   const tab_group_index = 1
   const source_data = {
@@ -110,15 +116,15 @@ function toDifferentWindow( t ) {
 
   const state1 = moveGroup( state0, { source_data, target_data } )
 
-  t.ok( validateState( state1 ), "state validates", validateState.errors )
+  t.ok( validateState( state1 ), "should pass validation", validateState.errors )
   t.equal( state1.windows[ 0 ].tab_groups.length, 2 )
   t.equal( state1.windows[ 1 ].tab_groups.length, 4 )
   t.equal( state1.windows[ 1 ].tab_groups[ tab_group_index ].id, tab_group_id )
 
   t.end()
-}
+})
 
-function activeToDifferentWindow( t ) {
+tap.test( function activeGroupToDifferentWindow( t ) {
   const tab_group_id = 2
   const tab_group_index = 1
   const source_data = {
@@ -156,16 +162,161 @@ function activeToDifferentWindow( t ) {
 
   const state1 = moveGroup( state0, { source_data, target_data } )
 
-  t.ok( validateState( state1 ), "state validates", validateState.errors )
+  t.ok( validateState( state1 ), "should pass validation", validateState.errors )
   t.equal( state1.windows[ 0 ].tab_groups.length, 2 )
   t.equal( state1.windows[ 0 ].active_tab_group_id, 3 )
   t.equal( state1.windows[ 1 ].tab_groups.length, 4 )
   t.equal( state1.windows[ 1 ].tab_groups[ tab_group_index ].id, tab_group_id )
 
   t.end()
-}
+})
 
-tap.test( toSameWindow )
-tap.test( toSameWindowDropZone )
-tap.test( toDifferentWindow )
-tap.test( activeToDifferentWindow )
+tap.test( function lastGroupToDifferentWindow( t ) {
+  const tab_group_index = 1
+  const source_data = {
+    window_id: 1,
+    tab_group_id: 3
+  }
+  const target_data = {
+    window_id: 2,
+    tab_group_index
+  }
+
+  const state0 = {
+    config: {},
+    windows: [
+      createWindow( source_data.window_id, [
+        createPinnedTabGroup( [] ),
+        createTabGroup( source_data.tab_group_id, [
+          createTestTab({ id: 6 })
+        ]),
+      ]),
+      createWindow( target_data.window_id, [
+        createPinnedTabGroup( [] ),
+        createTabGroup( 4, [
+          createTestTab({ id: 8 })
+        ]),
+      ])
+    ]
+  }
+
+  const state1 = moveGroup( state0, { source_data, target_data } )
+
+  t.ok( validateState( state1 ), "should pass validation", validateState.errors )
+  t.equal( state1.windows.length, 1, "should remove empty window" )
+
+  t.end()
+})
+
+tap.test( function lastGroupToDifferentWindowWithPinned( t ) {
+  const tab_group_index = 1
+  const source_data = {
+    window_id: 1,
+    tab_group_id: 3
+  }
+  const target_data = {
+    window_id: 2,
+    tab_group_index
+  }
+
+  const state0 = {
+    config: {},
+    windows: [
+      createWindow( source_data.window_id, [
+        createPinnedTabGroup( [
+          createTestTab({ id: 9 })
+        ] ),
+        createTabGroup( source_data.tab_group_id, [
+          createTestTab({ id: 6 })
+        ]),
+      ]),
+      createWindow( target_data.window_id, [
+        createPinnedTabGroup( [] ),
+        createTabGroup( 4, [
+          createTestTab({ id: 8 })
+        ]),
+      ])
+    ]
+  }
+
+  const state1 = moveGroup( state0, { source_data, target_data } )
+
+  t.ok( validateState( state1 ), "should pass validation", validateState.errors )
+  t.equal( state1.windows.length, 2, "should not remove source window" )
+  t.equal( state1.windows[ 0 ].tab_groups.length, 2, "should add a new group" )
+
+  t.end()
+})
+
+tap.test( function lastNonEmptyGroupToDifferentWindow( t ) {
+  // Latest firefox doesn't trigger window create first :(
+  const tab_group_index = 1
+  const source_data = {
+    window_id: 1,
+    tab_group_id: 3
+  }
+  const target_data = {
+    window_id: 2,
+    tab_group_index
+  }
+
+  const state0 = {
+    config: {},
+    windows: [
+      createWindow( source_data.window_id, [
+        createPinnedTabGroup( [] ),
+        createTabGroup( source_data.tab_group_id, [
+          createTestTab({ id: 6 })
+        ]),
+        createTabGroup( 5, [] ),
+      ]),
+      createWindow( target_data.window_id, [
+        createPinnedTabGroup( [] ),
+        createTabGroup( 4, [
+          createTestTab({ id: 8 })
+        ]),
+      ])
+    ]
+  }
+
+  const state1 = moveGroup( state0, { source_data, target_data } )
+
+  t.ok( validateState( state1 ), "should pass validation", validateState.errors )
+  t.equal( state1.windows.length, 1, "should remove empty window" )
+
+  t.end()
+})
+
+tap.test( function toNewWindow( t ) {
+  const tab_group_index = 1
+  const source_data = {
+    window_id: 1,
+    tab_group_id: 3
+  }
+  const target_data = {
+    window_id: 2, // This window doesn't exist yet
+    tab_group_index: 1,
+  }
+
+  const state0 = {
+    config: {},
+    windows: [
+      createWindow( source_data.window_id, [
+        createPinnedTabGroup( [] ),
+        createTabGroup( source_data.tab_group_id, [
+          createTestTab({ id: 6 })
+        ]),
+        createTabGroup( 5, [
+          createTestTab({ id: 7 })
+        ]),
+      ]),
+    ]
+  }
+
+  const state1 = moveGroup( state0, { source_data, target_data } )
+
+  t.ok( validateState( state1 ), "should pass validation", validateState.errors )
+  t.equal( state1.windows.length, 2, "should add a new window" )
+
+  t.end()
+})