+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.0.17] - 2020-03-22
+
+### Fixed
+- validation error on close last tab in window
+- validation error on moving the last tabs from a window
+- move active group between windows
+- handling when active group is closed
+- feature toggle hotkey in older browsers
+
 ## [0.0.16] - 2020-02-17
 
 ### Added
@@ -38,7 +47,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - header overlap for drag target position ink
 - new group icon display edge case
 - active tab display in blue
-- drag styling on hilighted tabs while group is moving
+- drag styling on highlighted tabs while group is moving
 - flash of drag styling on active tab switch
 
 ## [0.0.13] - 2019-07-21
@@ -93,7 +102,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - use sync for config settings
 - incremental display search results
-- use native hilighted tabs for multi-drag
+- use native highlighted tabs for multi-drag
 
 ### Fixed
 - scrollbar display on windows
@@ -178,6 +187,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - prevent closing last tab group in a window
 
 [Unreleased]: https://github.com/amccausl/tabulate/compare/master...develop
+[0.0.17]: https://github.com/amccausl/tabulate/compare/v0.0.16...v0.0.17
 [0.0.16]: https://github.com/amccausl/tabulate/compare/v0.0.15...v0.0.16
 [0.0.15]: https://github.com/amccausl/tabulate/compare/v0.0.14...v0.0.15
 [0.0.14]: https://github.com/amccausl/tabulate/compare/v0.0.13...v0.0.14